backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
pyodbc==5.0.1
pymssql==2.2.11
email-validator==2.1.0
werkzeug==3.0.1
